pyproject.toml
[project]
name = "job-scraper-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.187.0",
    "langchain-openai>=1.1.5",
    "openai>=2.13.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "supadata>=1.5.0",
]
